Orders list

+++ app/Http/Controllers/Api/Product/ProductController.php
@@ -104,7 +104,7 @@ class ProductController extends ApiController
         $data = $input
             ->string('productId')->validate('required|exists:products,id')->alias('product_id')->next()
             ->string('dateStart')->validate('required|date')->alias('date_start')->next()
-            ->integer('price')->validate('required')->next()
+            ->numeric('price')->validate('required')->next()
             ->integer('vatPercentage')->validate('required')->alias('vat_percentage')->next()
             ->get();
